✨ 拡張子をtsxにした。マーカー表示にした。
